Prepping for adding downloads to watchlist courses
