+++ src/components/MyCourses/MyCoursesSection.js
@@ -4,20 +4,46 @@ import Container from '@mui/material/Container'
 import Typography from '@mui/material/Typography'
 import Tabs from '@mui/material/Tabs'
 import Tab from '@mui/material/Tab'
+import Snackbar from '@mui/material/Snackbar'
+import MuiAlert from '@mui/material/Alert'
 
 import Section from '../Section'
 import MyCoursesProgress from './MyCoursesProgress'
 import MyCoursesDownloads from './MyCoursesDownloads'
 import MyCoursesWatchlistDrawer from './MyCoursesWatchlistDrawer'
 import { useTranslation } from 'react-i18next'
+import { useAuth } from '../../util/auth'
 
 function MyCoursesSection(props) {
   const { t } = useTranslation()
+  const auth = useAuth()
 
   const [tabIndex, setTabIndex] = useState(0)
+  const [courseToDownload, setCourseToDownload] = useState(false)
+  const [downloadVideos, setDownloadVideos] = useState(false)
+  const [openSnackbar, setOpenSnackbar] = useState(false)
+  const [snackbarMessage, setSnackbarMessage] = useState(
+    'Login to add to your watchlist',
+  )
 
   const handleChangeTab = (event, newTab) => {
     setTabIndex(newTab)
+  }
+
+  const Download = () => {
+    return (
+      <div style={{ display: 'none' }}>
+        {courseToDownload.map((video, index) =>
+          video?.link ? (
+            <iframe
+              key={`${video.link}-${index}`}
+              title={`${video.link}-${index}`}
+              src={video.link}
+            />
+          ) : null,
+        )}
+      </div>
+    )
   }
 
   return (
@@ -61,7 +87,12 @@ function MyCoursesSection(props) {
               </Tabs>
             </Box>
             <TabPanel tabIndex={tabIndex} index={0}>
-              <MyCoursesProgress />
+              <MyCoursesProgress
+                setCourseToDownload={setCourseToDownload}
+                setDownloadVideos={setDownloadVideos}
+                setSnackbarMessage={setSnackbarMessage}
+                setOpenSnackbar={setOpenSnackbar}
+              />
             </TabPanel>
             <TabPanel tabIndex={tabIndex} index={1}>
               <MyCoursesDownloads />
@@ -70,6 +101,22 @@ function MyCoursesSection(props) {
               <MyCoursesWatchlistDrawer />
             </TabPanel>
           </Box>
+          {downloadVideos && <Download />}
+          <Snackbar
+            open={openSnackbar}
+            autoHideDuration={6000}
+            onClose={() => setOpenSnackbar(false)}
+          >
+            <MuiAlert
+              elevation={6}
+              variant="filled"
+              onClose={() => setOpenSnackbar(false)}
+              severity={auth.user ? 'success' : 'warning'}
+              sx={{ width: '100%' }}
+            >
+              {snackbarMessage}
+            </MuiAlert>
+          </Snackbar>
         </Container>
       </Box>
     </Section>
