+++ src/components/MyCourses/handleAddToDownloads.js
@@ -1,8 +1,4 @@
-import {
-  downloadsData,
-  createDownloadCourse,
-  updateDownloads,
-} from '../../util/db'
+import { createDownloadCourse, updateDownloads } from '../../util/db'
 
 const getVideoDataToDownload = (videoInfo) => {
   return videoInfo.map((video) => {
@@ -45,19 +41,21 @@ export const handleAddToDownloads = (
     courseUID: course.uid,
     videos: videoDownloadData,
   }
-  const videosAlreadyAdded = downloadsData[0].videos
-  const videosToAdd = videoDownloadData.filter(
-    (video) =>
-      !videosAlreadyAdded.map((video) => video.uri).includes(video.uri),
-  )
+
   const isInDownloadsList = downloadsData.length > 0
-  const hasVideosToAdd = videosToAdd.length > 0
 
   if (!isInDownloadsList) {
     createDownloadCourse(downloadedCourse).then(() =>
       handleSnackbar('Success!  Added to your Downloads'),
     )
   } else {
+    const videosAlreadyAdded = downloadsData[0].videos
+    const videosToAdd = videoDownloadData.filter(
+      (video) =>
+        !videosAlreadyAdded.map((video) => video.uri).includes(video.uri),
+    )
+    const hasVideosToAdd = videosToAdd.length > 0
+
     if (hasVideosToAdd) {
       updateDownloads(downloadsData[0].id, {
         ...downloadsData[0],
